refactor: modify the folder structure
