+++ src/components/Header/Header.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
 import { PURPLE } from '../../assets/colors'
-import SearchBar from '../SearchBar'
-import Logo from '../Logo'
+import SearchBar from '../Header/SearchBar'
+import Logo from '../Header/Logo'
 
 const StyledHeader = styled.div`
   width: 100%;
